+++ src/Realm.Persistance/Data/Inventory.cs
@@ -4,7 +4,7 @@ public sealed class Inventory
 {
     public int Id { get; set; }
     public int UserId { get; set; }
-    public uint Size { get; set; }
+    public decimal Size { get; set; }
 
     public ICollection<InventoryItem> InventoryItems = new List<InventoryItem>();
 }
